requirements.txt
xmltodict
requests
urllib3
chardet
certifi
tinydb
lxml
fake_useragent
python-dotenv
